requirements.txt
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.1
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.56.0
ghostscript==0.7
iniconfig==2.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
pyparsing==3.2.1
pypdf==3.17.4
pypdfium2==4.30.1
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tabulate==0.9.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
